requirements.txt
urllib3
requests
beautifulsoup4
google-generativeai
python-dotenv
app-store-scraper